pyproject.toml
[project]
name = "ard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opencv-python>=4.12.0.88",
    "pyserial>=3.5",
    "ultralytics>=8.3.233",
]
